Split out setupMethod to wrapper.js
